move schematic_editor to plugins folder
