+++ gdsfactory/plugins/schematic_editor/__init__.py
@@ -1,0 +1,3 @@
+from gdsfactory.plugins.schematic_editor.schematic_editor import SchematicEditor
+
+__all__ = ["SchematicEditor"]
